Get sections and cache right, also added fill colors
We weren't using /cache:cache as mal.cache, so we were only
getting direct topic links. We also weren't getting topic
links for subsections of pages.

This commit adds fill colors for nodes. They're based on the
prefix of the ID (before the first -, or the whole ID), and
cycles through the light Tango colors.